+++ example/lib/main.dart
@@ -28,13 +28,16 @@ class _TabbedViewExamplePageState extends State<TabbedViewExamplePage> {
 
     tabs.add(TabData(
         text: 'Tab 1',
+        buttons: [TabButton(iconPath: TabbedViewIcons.close,onPressed: (){})],
         content: Padding(child: Text('Hello'), padding: EdgeInsets.all(8))));
+    for(int i = 2; i< 8;i++) {
+      tabs.add(TabData(
+          text: 'Tab $i',
+          content:
+          Padding(child: Text('Hello again'), padding: EdgeInsets.all(8))));
+    }
     tabs.add(TabData(
-        text: 'Tab 2',
-        content:
-            Padding(child: Text('Hello again'), padding: EdgeInsets.all(8))));
-    tabs.add(TabData(
-        text: 'Tab 3',
+        text: 'TextField',
         content: Padding(
             child: TextField(
                 decoration: InputDecoration(
@@ -47,8 +50,14 @@ class _TabbedViewExamplePageState extends State<TabbedViewExamplePage> {
 
   @override
   Widget build(BuildContext context) {
-    TabbedView tabbedView = TabbedView(controller: _controller);
+    TabbedView tabbedView = TabbedView(controller: _controller,tabsAreaButtonsBuilder: _tabsAreaButtonsBuilder);
+    TabbedViewTheme theme = TabbedViewTheme(child: tabbedView, data: TabbedViewThemeData.mobile());
     return Scaffold(
-        body: Container(child: tabbedView, padding: EdgeInsets.all(32)));
+        body: Container(child: theme, padding: EdgeInsets.all(32)));
+  }
+
+  List<TabButton> _tabsAreaButtonsBuilder(
+      BuildContext context, int tabsCount){
+    return [TabButton(iconPath: TabbedViewIcons.close,onPressed: (){})];
   }
 }