+++ lib/src/tab_button.dart
@@ -1,19 +1,27 @@
 import 'package:flutter/widgets.dart';
+import 'package:tabbed_view/src/icon_path.dart';
 import 'package:tabbed_view/src/tabbed_view_menu_builder.dart';
 
 /// Configures a tab button.
 class TabButton {
   TabButton(
-      {required this.icon,
+      {this.iconData,
+      this.iconPath,
       this.color,
       this.hoverColor,
       this.disabledColor,
       this.onPressed,
       this.menuBuilder,
       this.toolTip,
-      this.padding});
+      this.padding}) {
+    if ((this.iconData != null && this.iconPath != null) ||
+        (this.iconData == null && this.iconPath == null)) {
+      throw ArgumentError('iconData or iconPath must be set');
+    }
+  }
 
-  final IconData icon;
+  final IconData? iconData;
+  final IconPath? iconPath;
   final Color? color;
   final EdgeInsetsGeometry? padding;
   final Color? hoverColor;